+++ EVCP.MapLoader/OsmJsonParser.cs
@@ -5,6 +5,7 @@ using System.Numerics;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
+using EVCP.Domain.Models;
 
 namespace EVCP.MapLoader;
