+++ EVCP.MapLoader/OsmXmlParser.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Numerics;
 using System.Xml.Linq;
-
+using EVCP.Domain.Models;
 
 namespace EVCP.MapLoader;
